hooked up Firestore and connected profile to database; todo - associate data with logged in users

+++ Adjacent Finances/Models/UserProfileDataModel.swift
@@ -10,6 +10,6 @@ import Foundation
 import Firebase
 
 struct user {
-    let savingsAmount: Int
+    let savingsAmount: Double
     let birthYear: Int
 }
